+++ test/unit/Tests/IO.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE OverloadedLists #-}
+
 -- | IO Unit tests.
 --
 -- @since 0.1.0.0
@@ -5,7 +7,10 @@ module Tests.IO (tests) where
 
 import Control.Exception (try)
 import Data.Version.Package qualified as PV
-import Data.Version.Package.Internal (PackageVersion (..), ReadFileError (..))
+import Data.Version.Package.Internal
+  ( PackageVersion (UnsafePackageVersion),
+    ReadFileError (ReadFileErrorGeneral),
+  )
 import Test.Tasty (TestTree)
 import Test.Tasty qualified as Tasty
 import Test.Tasty.HUnit ((@=?))