+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
 ## [0.4]
 ### Removed
 * `Pretty` instances as these were not very useful.
+* Internal representation now uses `NonEmpty` instead of `[]`.
+* `MkPackageVersion` and `unPackageVersion` use `NonEmpty`.
 
 ### Added
 * Updated bounds on `base`, `bytestring`, and `template-haskell`.